pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "nteja13",email = "your.nteja13@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "supabase (>=2.28.3,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
